+++ DefaultNodes/Exec/GateNode.cs
@@ -19,12 +19,14 @@ namespace Exodrifter.NodeGraph.DefaultNodes
 		internal ExecType execIn = ExecType.None;
 
 		[Description("True if the gate is open.")]
-		[SerializeField, Input("IsOpen", 20, SocketFlags.Editable)]
+		[SerializeField, Input("IsOpen", SocketFlags.Editable)]
 		internal bool isOpen = false;
 
 		[Description("The execution signal that passed through the gate.")]
 		[SerializeField, Output("Exec Out")]
 		internal ExecType execOut = ExecType.None;
+
+		public override float InputWidth { get { return 30; } }
 
 		#region Methods